grab first paragraph from Contentful for Joyent case study (instead of intro sentence) (#78)
## Change existing Joyent case study - currently pulling intro sentence when it shouldn't - [Trello ticket](https://trello.com/c/wgVGuD0M/223-change-existing-joyent-case-study-currently-pulling-intro-sentence-when-it-shouldnt)

### Background info
![image](https://user-images.githubusercontent.com/15656538/50840871-5de8cb80-135b-11e9-9602-a189687ef0b2.png)

The first paragraph on the Joyent case study ("In the early fall of..."), was entered into Contentful as `introSentence`. So, in the codebase, we are fetching the intro sentence for this page.

But this is incorrect. The `introSentence` field should be used for the "featured case study" component, or the case study cards that are going on the Our work page. Whereas, "In the early fall of..." should be entered into Contentful as `genericText1`

### This PR
So I've changed the content in Contentful, published (but not yet deployed).

This PR contains the necessary change to the codebase, in order to reflect the content alterations